+++ paper/paper.bib
@@ -11,7 +11,7 @@
  urldate = {2022-07-19},
  volume = {266},
  year = {1992},
- doi = {10.2307/24938980}
+ doi = {10.1038/scientificamerican0392-44}
 }
 
 @article{Komendantov:2019,
@@ -59,7 +59,7 @@
   pages={aac9462},
   year={2015},
   publisher={American Association for the Advancement of Science},
-  doi = {10.1126/science.aac946}
+  doi = {10.1126/science.aac9462}
 }
 
 @article{Clements:1997,
