### 2022-07-27 #4613 @Kevin-Mattheus-Moerman
@jmschrei this may be off topic a bit but let me know if you can handle this work featuring signal processing for a neuroscience application (several editors closer to this domain are currently handling many other submissions).
