+++ docs/resources.rst
@@ -1,12 +1,12 @@
 Other resources
 ==================
-List of availabe tools/resources related to Patch-seq experiments. I only listed open-source tools/resources (and 
+List of available tools/resources related to Patch-seq experiments. I only listed open-source tools/resources (and 
 github repo links if any). 
-If you would (not) like to list your *open-source* tools/resources here, please post a request in github issue page
+If you would like to list your **open-source** tools/resources here, please post a request in github issue page
 
 Whole-cell recording analysis
 -----------------------------------
-- `eventer <https://github.com/acp29/eventer>`_  "Software for the detection of spontaneous synaptic currents" in **Matlab**
+- `eventer <https://github.com/acp29/eventer>`_  "Software for the detection of spontaneous synaptic currents" in *Matlab*
 - `Stimfit <https://github.com/neurodroid/stimfit>`_  "Stimfit is a free, fast and simple program for viewing and analyzing electrophysiological data"
 - `aisynphys <https://github.com/campagnola/aisynphys>`_ "Analysis tools specific to the aiephys multipatch project."
 - `pyABF <https://github.com/swharden/pyABF>`_  "pyABF is a Python library for reading electrophysiology data from Axon Binary Format (ABF) files."
@@ -15,5 +15,9 @@ Whole-cell recording analysis
 Morphology analysis
 -----------------------------------
 - `NeuroM <https://github.com/BlueBrain/NeuroM>`_  "NeuroM is a Python toolkit for the analysis and processing of neuron morphologies."
-- `TREES toolbox <https://github.com/cuntzlab/treestoolbox>`_  Comprehensive morphology analysis and modeling in **Matlab**
+- `TREES toolbox <https://github.com/cuntzlab/treestoolbox>`_  Comprehensive morphology analysis and modeling in *Matlab*
 
+Data standard
+-----------------
+- `Neurodata Without Borders <https://www.nwb.org/>`_ "Neurodata Without Borders (NWB) is a data standard for neurophysiology, providing neuroscientists with a common standard to share, archive, use, and build analysis tools for neurophysiology data."
+- `PyNWB <https://github.com/NeurodataWithoutBorders/pynwb>`_ "PyNWB is a Python package for working with NWB files. "
