### 2022-08-24 #4706 @editorialbot
[REVIEW]: PatchView: A Python Package for Patch-clamp Data Analysis and Visualization
**Submitting author:** @ZeitgeberH (<a href="http://orcid.org/0000-0002-5254-5315"></a>)
**Repository:** https://github.com/ZeitgeberH/patchview
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.6
