+++ patchview/patchview.py
@@ -37,6 +37,7 @@ from patchview.utilitis.AnalysisMethods import (
     cleanASCfile,
 )
 from patchview.utilitis import fitFuncs
+from patchview.utilitis.morphorFeatureExtrator import getSomaStats, extractMorhporFeatures
 from patchview.utilitis.patchviewObjects import *
 import networkx as nx
 
@@ -658,7 +659,7 @@ class MainWindow(QtWidgets.QMainWindow):
                 "maximal radius": [],
                 "minimal radius": [],
                 "min_max_ratio": [],
-                "maximal diamter": [],
+                "maximal diameter": [],
                 "intercell distance": [],
             }
 
@@ -666,19 +667,17 @@ class MainWindow(QtWidgets.QMainWindow):
                 fig2D, ax2D = morphor_viewer.draw(
                     n, mode="2d", fig=fig2D, ax=ax2D, label=str(idx + 1)
                 )
-                # morp, ax2Dhor_viewer.draw(n, mode='3d',fig=fig3D, ax =ax3D)
-                ps = n.soma.points[:, :3]
-                maxDia = self.minmaxDist(ps)
-                ps_center = np.nanmean(ps, axis=0)
-                ps_radius = np.sqrt(np.nansum((ps - ps_center) ** 2, axis=1))
-                ps_avgRadius = np.nanmean(ps_radius)
-                ps_center = np.nanmean(ps, axis=0)
-
-                centers["X"].append(ps_center[0])
-                centers["Y"].append(ps_center[1])
-                centers["Z"].append(ps_center[2])
+                maxDia, soma_center, soma_radius, soma_avgRadius = getSomaStats(n)
+                centers["X"].append(soma_center[0])
+                centers["Y"].append(soma_center[1])
+                centers["Z"].append(soma_center[2])
+                centers["average radius"].append(soma_avgRadius)
+                centers["maximal radius"].append(np.max(soma_radius))
+                centers["minimal radius"].append(np.min(soma_radius))
+                centers["min_max_ratio"].append(np.min(soma_radius) / np.max(soma_radius))
+                centers["maximal diameter"].append(maxDia)
                 distanceList_ = []
-                if len(centers["X"]) > 1:
+                if len(centers["X"]) > 1:## calculte distance between cells 
                     x0, y0, z0 = centers["X"][-1], centers["Y"][-1], centers["Z"][-1]
                     for kkk, _ in enumerate(centers["X"][:-1]):
                         d = np.round(
@@ -690,12 +689,6 @@ class MainWindow(QtWidgets.QMainWindow):
                             1,
                         )
                         distanceList_.append(d)
-
-                centers["average radius"].append(ps_avgRadius)
-                centers["maximal radius"].append(np.max(ps_radius))
-                centers["minimal radius"].append(np.min(ps_radius))
-                centers["min_max_ratio"].append(np.min(ps_radius) / np.max(ps_radius))
-                centers["maximal diamter"].append(maxDia)
                 centers["intercell distance"].append(distanceList_)
                 centers["Name"].append(str(idx + 1))
 
@@ -762,30 +755,9 @@ class MainWindow(QtWidgets.QMainWindow):
                 "number_of_sections_per_neurite", neurons
             )
             df_summary = {}
-            ps = neurons.soma.points[:, :2]
-            maxDia = self.minmaxDist(ps)
-            ps_center = np.mean(ps, axis=0)
-            ps_radius = np.sqrt(np.sum((ps - ps_center) ** 2, axis=1))
-            ps_avgRadius = np.mean(ps_radius)
             df_summary["ASC file"] = [fname]
[…]
