+++ setup.py
@@ -44,12 +44,13 @@ setup(
     include_package_data=True,
     keywords='patchview',
     name='patchview',
-    packages=find_packages(include=['patchview','patchview/*']),
+    packages=find_packages(include=['patchview','patchview/Data',\
+        'patchview/HekaIO','patchview/neurom','patchview/utilitis']),
     setup_requires=setup_requirements,
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/zeitgeberH/patchview',
-    version='0.2.2',
+    version='0.2.5',
     zip_safe=False,
     entry_points={
         'gui_scripts': [
