+++ requirements.txt
@@ -1,14 +1,14 @@
 colorcet
-pynwb
 Pillow
+pynwb
 tqdm
-PySide2==5.15.2 
+PySide2
 pyqt5
 pyqtwebengine
 pyqt5-sip
-pyqtgraph==0.12.4
-PyOpenGL==3.1.6
-PyOpenGL_accelerate==3.1.6
+pyqtgraph
+PyOpenGL
+PyOpenGL_accelerate>=3.1.6
 cython
 neo
 pylru==1.2.1
@@ -17,8 +17,11 @@ joblib
 future==0.18.2
 pyYAML==6.0
 python-dateutil
+seaborn
 pytz
 scipy
 pandas
 matplotlib
 scikit-image
+sklearn
+numpy==1.21.5
