+++ enviroment.yml
@@ -3,7 +3,7 @@ channels:
   - defaults
 dependencies:
   - pip=21.0.1
-  - python=3.8
+  - python>=3.8
   - pip:
     - pandas
     - matplotlib
@@ -12,9 +12,9 @@ dependencies:
     - Pillow
     - tqdm
     - pyqtwebengine
-    - PySide2==5.15.2
-    - PyQT5==5.12
-    - pyqtgraph==0.12.4
+    - PySide2
+    - PyQT5
+    - pyqtgraph
     - PyOpenGL
     - PyOpenGL_accelerate
     - cython
