+++ patchview/neurom/stats.py
@@ -1,211 +1,0 @@
-# Copyright (c) 2015, Ecole Polytechnique Federale de Lausanne, Blue Brain Project
-# All rights reserved.
-#
-# This file is part of NeuroM <https://github.com/BlueBrain/NeuroM>
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-#     1. Redistributions of source code must retain the above copyright
-#        notice, this list of conditions and the following disclaimer.
-#     2. Redistributions in binary form must reproduce the above copyright
-#        notice, this list of conditions and the following disclaimer in the
-#        documentation and/or other materials provided with the distribution.
-#     3. Neither the name of the copyright holder nor the names of
-#        its contributors may be used to endorse or promote products
-#        derived from this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-# ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY
-# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-# ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-"""Statistical analysis helper functions
-
-Nothing fancy. Just commonly used functions using scipy functionality."""
-
-from collections import namedtuple
-from enum import Enum, unique
-
-import numpy as np
-from future.moves.collections import OrderedDict
-from scipy import stats as _st
-
-FitResults = namedtuple("FitResults", ["params", "errs", "type"])
-
-
-@unique
-class StatTests(Enum):
-    """Enum representing valid statistical tests of scipy"""
-
-    ks = 1
-    wilcoxon = 2
-    ttest = 3
-
-
-def get_test(stest):
-    """Returns the correct stat test"""
-    sts = {
-        StatTests.ks: "ks_2samp",
-        StatTests.wilcoxon: "wilcoxon",
-        StatTests.ttest: "ttest_ind",
-    }
-
-    if stest in StatTests:
-        return sts[stest]
-    raise TypeError("Statistical test not recognized. Choose from ks, wilcoxon, ttest.")
-
-
-def fit_results_to_dict(fit_results, min_bound=None, max_bound=None):
-    """Create a JSON-comparable dict from a FitResults object
-
-    Parameters:
-        fit_results (FitResults): object containing fit parameters,\
-            errors and type
-        min_bound: optional min value to add to dictionary if min isn't\
-            a fit parameter.
-        max_bound: optional max value to add to dictionary if max isn't\
-            a fit parameter.
-
-    Returns:
[…]
