+++ setup.py
@@ -15,6 +15,10 @@ requirements = []
 with open('requirements.txt', 'r') as f:
    requirements += [l for l in f.read().split('\n') if l.strip()]
 
+## forked version of NeuroM   
+requirements +=[
+    'NeuroM @ git+https://github.com/ZeitgeberH/NeuroM@patchview#egg=NeuroM'
+]
 setup_requirements = ['pytest-runner']
 
 test_requirements = ['pytest']

[…]
