+++ patchview/__init__.py
@@ -1,9 +1,9 @@
 from . import patchview
+from . import __version__
 from PyQt5 import QtWidgets
 import sys
 
-__version__ = "0.2.2"
-print("initilizating PatchView. use patchview.pvGUI() to launch the GUI")
+print(f"initilizating PatchView (version {__version__.version}).\nuse patchview.pvGUI() to launch the GUI")
 pvApp = QtWidgets.QApplication(sys.argv)
 
 def pvGUI():
