+++ patchview/patchview.py
@@ -1,5 +1,5 @@
 """
-Patchviewer by M.H.
+Patchview by M.H.
 """
 from PyQt5 import QtGui, QtWidgets, QtCore
 from PyQt5.QtWidgets import QMessageBox, QTableWidgetItem
@@ -41,6 +41,7 @@ from patchview.utilitis.patchviewObjects import *
 import networkx as nx
 
 ## reading neuroluscida file
+from morphio import SomaType
 import neurom as morphor_nm
 from neurom import viewer as morphor_viewer
 from neurom.io.multiSomas import MultiSoma
@@ -63,14 +64,12 @@ from hdmf.backends.hdf5 import HDF5IO as _HDF5IO
 from hdmf.validate import ValidatorMap
 from hdmf.build import BuildManager, TypeMap
 import hdmf.common
-from hdmf.spec import NamespaceCatalog  # noqa: E402
-from hdmf.utils import docval, getargs, call_docval_func, get_docval, fmt_docval_args  # noqa: E402
-from hdmf.build import BuildManager, TypeMap  # noqa: E402
+from hdmf.spec import NamespaceCatalog  
+from hdmf.utils import docval, getargs, call_docval_func, get_docval, fmt_docval_args  
+from hdmf.build import BuildManager, TypeMap
 warnings.filterwarnings("ignore")
 
 patchview_dir, this_filename = os.path.split(__file__)
-__version__ = "0.2.5"
-
 class MainWindow(QtWidgets.QMainWindow):
     """
     Main frame.
@@ -414,10 +413,14 @@ class MainWindow(QtWidgets.QMainWindow):
         splitViewTab_morph.addMatPlotviewAtTop(["2D"], size=(100, 100), dpi=1000)
 
         splitViewTab_morph.addTablesAtBottomRight(
-            ["Summary", "Distance (um)", "Axon", "Apic dendrite", "Basal dendrite"],
+            ["Summary", "Distance (um)"],
             editable=True,
             sortable=False,
         )
+        self.morphAnaFigs_matplotView = MatplotView() ## host for morph analysis figures
+        splitViewTab_morph.bottomRight_tabview.addTab(self.morphAnaFigs_matplotView,"Figures")
+
+        
         splitViewTab_morph.addParameterToLeftTab(
             "Analysis", AllMyPars.Morphor_analysis, self.morph_analysis_event
         )
@@ -646,10 +649,14 @@ class MainWindow(QtWidgets.QMainWindow):
             drawContour = True
         else:
             drawContour = False
+        if pv["Options"][1]["Ignore diameters"][0]:
+            realDia = True
+        else:
+            realDia= False
         self.splitViewTab_morph.matplotViews["2D"].getFigure().set_size_inches(
             5, 5, forward=False
         )
-        neurons = morphor_nm.load_morphology(fname)
+        neurons = morphor_nm.load_morphology(fname, somaType = SomaType.SOMA_CYLINDERS)
         fig2D = self.splitViewTab_morph.matplotViews["2D"].getFigure()
         fig2D.clf()
         ax2D = fig2D.add_subplot(111)
@@ -723,8 +730,8 @@ class MainWindow(QtWidgets.QMainWindow):
         else:  ## single neuron with dendtrite and/or axons
 
             morphor_viewer.draw(
-                neurons, mode="2d", realistic_diameters=True, fig=fig2D, ax=ax2D,
-                contour_on=drawContour, contour_color='g', contour_linewidth=0.2
+                neurons, mode="2d", realistic_diameters=realDia, fig=fig2D, ax=ax2D,
+                contour_on=drawContour, contour_color='g', contour_linewidth=0.2,
             )
[…]
