+++ tox.ini
@@ -1,11 +1,11 @@
 [tox]
-envlist = py{36,37,38}, flake8
+envlist = py{38,39,310}, flake8
 
 [travis]
 python =
+    3.10: py310
+    3.9: py39
     3.8: py38
-    3.7: py37
-    3.6: py36
 
 [testenv:flake8]
 basepython = python
