+++ setup.py
@@ -27,13 +27,15 @@ setup(
     author="Ming Hu",
     author_email='',
     classifiers=[
-        'Development Status :: 2 - Pre-Alpha',
-        'Intended Audience :: patch-clamp users',
-        'License :: OSI Approved :: BSD 3-Clause License',
+        'Development Status :: 5 - Production/Stable',
+        'Topic :: Scientific/Engineering :: Information Analysis',
+        'Intended Audience :: End Users/Desktop',
+        'License :: OSI Approved :: BSD License',
         'Natural Language :: English',
         "Programming Language :: Python :: 3",
         'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
+        'Programming Language :: Python :: 3.10',
         "Operating System :: OS Independent",
     ],
     description="Patchview perform data analysis and visualization on whole-cell recording data, including firing pattern analysis, event analysis, synatpic connection detection, morphorlocial analysis and more.",
