+++ docs/conf.py
@@ -55,9 +55,9 @@ author = u"Ming Hu"
 # the built documents.
 #
 # The short X.Y version.
-version =  "0.2.2"
+version =  "0.2.5"
 # The full version, including alpha/beta/rc tags.
-release =  "0.2.2"
+release =  "0.2.5"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
