+++ enviroment-dev.yml
@@ -1,26 +1,0 @@
-name: patchviewPy3
-channels:
-  - defaults
-dependencies:
-  - pip=21.0.1
-  - python=3.8
-  - pip:
-    - pandas
-    - matplotlib
-    - scikit-image
-    - colorcet
-    - Pillow
-    - tqdm
-    - pyqtwebengine
-    - PySide2
-    - PyQT5
-    - pyqtgraph
-    - PyOpenGL
-    - PyOpenGL_accelerate
-    - spyder
-    - cython
-    - pyinstaller
-    - joblib
-    - neo
-    - pylru==1.2.1
-    - h5py==3.7.0
