+++ docs/conf.py
@@ -21,6 +21,8 @@
 import os
 import sys
 sys.path.insert(0, os.path.abspath('..'))
+import patchview.__version__
+
 from typing import Any, Dict
 
 # -- General configuration ---------------------------------------------
@@ -55,9 +57,9 @@ author = u"Ming Hu"
 # the built documents.
 #
 # The short X.Y version.
-version =  "0.2.2"
+version =  patchview.__version__.version
 # The full version, including alpha/beta/rc tags.
-release =  "0.2.2"
+release =  patchview.__version__.version
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
