+++ docs/tutorials.rst
@@ -136,7 +136,7 @@ in stead of the plot widget's build-in save button.
     :width: 800
     :alt: Alternative text
 
-click "sholl analysis" to perform Sholl analysis from NeuroM package.
+Click "Sholl analysis" to perform "Sholl analysis" from NeuroM package.
 
 .. image:: resources/images/shollAna1.png
     :width: 800
