+++ requirements.txt
@@ -24,4 +24,4 @@ pandas
 matplotlib
 scikit-image
 sklearn
-numpy==1.21.5
+numpy>=1.22,<1.24
