+++ enviroment-dev.yml
@@ -24,3 +24,4 @@ dependencies:
     - neo
     - pylru==1.2.1
     - h5py==3.7.0
+    - numpy>=1.22,<1.24
