+++ setup.py
@@ -52,7 +52,7 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/zeitgeberH/patchview',
-    version='0.2.5',
+    version='0.2.5.1',
     zip_safe=False,
     entry_points={
         'gui_scripts': [
