### 2022-09-17 #4706 @ZeitgeberH
> Hi,
>
> this looks great! Fantastic piece of software. I have left a few Issues and PRs on the original repository. Mostly those are minor, so if @ZeitgeberH resolves them (or closes with justification), I think this is good to go.
>
> Thanks, Jan

Hi Jan,
Thanks very much for your suggestions. I've made changes accordingly and closed the issues.
Zeta
