### 2022-09-28 ZeitgeberH/patchview#7 @mstimberg
[JOSS review] Clarify/fix some menu items
I think users would benefit from a review of the menu and toolbar items, which are currently a bit inconsistent and sometimes confusing.

A few observations:
* The "File" menu has a clickable "---" separator, this should probably use `addSeparator` instead?
* The "Manual" entry in the "Help" menu does not seem to do anything
* Using an item such as "Analysis" → "Event analysis" will show an error box if the necessary data has not been loaded/selected, but others will do nothing and print an error to the console (e.g. "Analysis" → "Gap junction (Current tree). Maybe items that cannot be used should be grayed out and unselectable?
* The "Backgroud color" entry (also note the typo) should probably be called "switch Background color"?

openjournals/joss-reviews#4706
