### 2022-09-28 ZeitgeberH/patchview#8 @mstimberg
[JOSS review] test_readNwb fails on first run
The test suite checks whether reading the `test.nwb` file works, but this file is created by `test_writeNwb` which does not necessarily run before it (at least on my machine, tests run in alphabetical order).

openjournals/joss-reviews#4706
