### 2022-09-28 ZeitgeberH/patchview#9 @mstimberg
[JOSS review] cannot test connection analysis
As far as I can see, there is no multi-cell test data, so I don't have a way to test the connection analysis part (maybe generate some artificial data?).

openjournals/joss-reviews#4706
