### 2022-09-28 ZeitgeberH/patchview#10 @mstimberg
[JOSS review] Some morphology functionality is not available
Loading the `test.asc` morphology, I am not able to use "Sholl analysis" (prints "To be done!" on the command line), and "Update Cell names" and "Distance to Pia" do not seem to work either (errors on the command line, see below). I am also not quite sure how to test the [Plotting somas for a group of neurons](https://patchview.readthedocs.io/en/latest/introduction.html#plotting-somas-for-a-group-of-neurons) functionality – does this require a single file with multiple cells in it? I tried loading the same cell several times, but this leads to a broken display.

Error for "Update Cell names" (after loading a cell three times):
```pytb
Traceback (most recent call last):
  File "/mnt/data/miniforge3/envs/patchview/lib/python3.8/site-packages/patchview/patchview.py", line 6803, in morph_analysis_event
    self.updateInterCellDistance()
  File "/mnt/data/miniforge3/envs/patchview/lib/python3.8/site-packages/patchview/patchview.py", line 6747, in updateInterCellDistance
    summary.item(j, 9).value[p]
IndexError: invalid index to scalar variable.
```
Error for "Distance to Pia":
```pytb
Traceback (most recent call last):
  File "/mnt/data/miniforge3/envs/patchview/lib/python3.8/site-packages/patchview/patchview.py", line 6805, in morph_analysis_event
    self.measureDist2Pia()
  File "/mnt/data/miniforge3/envs/patchview/lib/python3.8/site-packages/patchview/patchview.py", line 6765, in measureDist2Pia
    if self.pia is not None:
AttributeError: 'MainWindow' object has no attribute 'pia'
```

openjournals/joss-reviews#4706
