### 2022-09-28 ZeitgeberH/patchview#11 @mstimberg
[JOSS review] Consider reading configuration file from standard directory
The configuration currently needs to be changed in the installation directory, which can be problematic (e.g. a system-wide installation) since the user does not necessarily has write privileges for this directory. A better solution would be to check for a configuration file in a standard configuration directory (e.g. by using https://github.com/ActiveState/appdirs).

openjournals/joss-reviews#4706
