### 2022-09-28 ZeitgeberH/patchview#12 @mstimberg
[JOSS Review] Some suggestions for GUI improvements
The GUI currently uses a non-standard interface, that can be sometimes a bit confusing (also see #7). In particular, some functionality is accessible by the menu, some by using the toolbar, and some by right-clicking. I think unifying this a bit could make things more straightforward. A few other remarks/observations (some certainly a matter of taste, please feel free to ignore them if you disagree, of course):
* I do not quite understand the use of separators in the toolbar – are these placeholders for future functionality? Or just a way to logically group elements that belong together?
* I was a bit confused by the fact that "Fire pattern" appears as:
  1. A menu item (Analysis → Fire pattern analysis)
  2. A menu option (Analysis → Fire pattern batch analysis)
  4. A tab on the bottom left and on the top
  5. A context-menu item in the tab on the bottom left ("Firing patterns (batch)")
  6. An entry in the toolbar ("Firing pattern")
→ Maybe this could be reduced/unified/clarified a bit?
* At a first glance, I did not immediately get the meaning of the separate "help tabs" that explain parameter settings (e.g. "Spike detection help" for "Spike detection" parameters), and they are not very readable, I think. Maybe you could consider adding the information as tooltips to the parameters directly?
* The GUI contains a few typos that might be worth being corrected. For example, "Vislization" → "Visualization" and "Backgroud color" → "Background color" in the menu; "replacem value" → "replacement value" in the Event Detection/preprocessing parameters

openjournals/joss-reviews#4706
