### 2022-09-28 #4706 @mstimberg
Hi. Apologies for the delay in reviewing the software in detail. I agree with @janfreyberg that this is a very useful software, providing access to many different tools in a single interface. I encountered a few issues, though, mostly around advertised functionality that I cannot test (ZeitgeberH/patchview#8, ZeitgeberH/patchview#9, ZeitgeberH/patchview#10). There are also a few things in the interface that might be worth improving (ZeitgeberH/patchview#7, ZeitgeberH/patchview#12), and I made a suggestion concerning the configuration file management (ZeitgeberH/patchview#11). The latter changes are rather "cosmetic", but I'd prefer to be able to test all the main functionality before giving my final stamp of approval. Thanks!
