+++ patchview/patchview.py
@@ -37,7 +37,8 @@ from patchview.utilitis.AnalysisMethods import (
     cleanASCfile,
 )
 from patchview.utilitis import fitFuncs
-from patchview.utilitis.morphorFeatureExtrator import getSomaStats, extractMorhporFeatures
+from patchview.utilitis.morphorFeatureExtrator import (
+    getSomaStats, extractMorhporFeatures, sholl_analysis)
 from patchview.utilitis.patchviewObjects import *
 import networkx as nx
 
@@ -412,10 +413,14 @@ class MainWindow(QtWidgets.QMainWindow):
         splitViewTab_morph.addMatPlotviewAtTop(["2D"], size=(100, 100), dpi=1000)
 
         splitViewTab_morph.addTablesAtBottomRight(
-            ["Summary", "Distance (um)", "Axon", "Apic dendrite", "Basal dendrite"],
+            ["Summary", "Distance (um)"],
             editable=True,
             sortable=False,
         )
+
+        self.morphAnaFigs_matplotView = MatplotView() ## host for morph analysis figures
+        splitViewTab_morph.bottomRight_tabview.addTab(self.morphAnaFigs_matplotView,"Figures")
+
         splitViewTab_morph.addParameterToLeftTab(
             "Analysis", AllMyPars.Morphor_analysis, self.morph_analysis_event
         )
@@ -643,6 +648,7 @@ class MainWindow(QtWidgets.QMainWindow):
             5, 5, forward=False
         )
         neurons, custom_data = morphor_nm.load_neuron(fname)
+        self.pia = None
         fig2D = self.splitViewTab_morph.matplotViews["2D"].getFigure()
         fig2D.clf()
         ax2D = fig2D.add_subplot(111)
@@ -695,7 +701,6 @@ class MainWindow(QtWidgets.QMainWindow):
             self.splitViewTab_morph.matplotViews["2D"].draw()
             xlim1 = ax2D.xaxis.get_data_interval().copy()
             ylim1 = ax2D.yaxis.get_data_interval().copy()
-
             if len(custom_data) > 0:
                 for datablock in custom_data:
                     d = datablock.data_block
@@ -722,11 +727,12 @@ class MainWindow(QtWidgets.QMainWindow):
             self.splitViewTab_morph.tables["Summary"].appendData(df)
             self.splitViewTab_morph.tables["Summary"].show()
             self.updateInterCellDistance()
-
+            self.neuronMorph =  None ## no dendrites for further analysis
         else:  ## single neuron with dendtrite and/or axons
             morphor_viewer.draw(
                 neurons, mode="2d", realistic_diameters=True, fig=fig2D, ax=ax2D
             )
+            self.neuronMorph =  neurons ## store dendrites for further analysis
             ax2D.axis("equal")  ## only works for 2D
             # ax2D.view_init(azim=0, elev=90)
             ax2D.set_title("{0}".format(fname), fontsize=12, color="k")
@@ -781,6 +787,28 @@ class MainWindow(QtWidgets.QMainWindow):
                 if diameter > maxDia:
                     maxDia = diameter
         return maxDia
+
+    def update_sholl(self):
+        if self.neuronMorph is not None:
+            sholl_dist, sholl_bins = sholl_analysis(self.neuronMorph)
+            
+            fig = self.morphAnaFigs_matplotView.getFigure()
+            fig.clf()
+            ax = fig.add_subplot(111)
+            ax.cla()
+            # ax.plot(sholl_bins, sholl_dist)
+            sns.set_style("whitegrid")
+            ax = sns.lineplot(x=sholl_bins, y=sholl_dist, ax=ax)
[…]
