### 2022-10-01 f1208bf
using appdirs for config file
