### 2022-10-01 098a2a1
documentation on morph analysis
