+++ docs/installation.rst
@@ -142,7 +142,11 @@ Configuration
 PatchView use a Yaml file for its basic configuration.
 
 * If you use the app version, it is located in the **patchview\\Data\\patchview.yaml** of the app folder.
-* If you install it via Pip or from source, it is located in **PATH-OF-YOUR-ENVIROMENT\\Lib\\site-packages\\Patchview-xx-py3.x.egg\\patchview\\Data\\patchview.yaml**.
+* If you install it via Pip or from source, it is located in ``user_config_dir`` generated by **appdirs**,
+  
+  For Windows: it is **'C:\\Users\\YOUR-USER-NAME\\AppData\\Local\\Patchview\\Patchview'**.
+
+  For Linux: it is **'/home/YOUR-USER-NAME/.config/Patchview'**
 
 Open **patchview.yaml** with any text editor.  
 
