+++ docs/installation.rst
@@ -142,11 +142,13 @@ Configuration
 PatchView use a Yaml file for its basic configuration.
 
 * If you use the app version, it is located in the **patchview\\Data\\patchview.yaml** of the app folder.
-* If you install it via Pip or from source, it is located in ``user_config_dir`` generated by **appdirs**,
+* If you install it via Pip or from source, it is located in ``user_config_dir`` generated by `appdirs`_,
   
   For Windows: it is **'C:\\Users\\YOUR-USER-NAME\\AppData\\Local\\Patchview\\Patchview'**.
 
   For Linux: it is **'/home/YOUR-USER-NAME/.config/Patchview'**
+
+.. _appdirs: https://github.com/ActiveState/appdirs
 
 Open **patchview.yaml** with any text editor.  
 
