### 2022-10-01 ZeitgeberH/patchview#9 @ZeitgeberH
A folder of test files recorded from a multi-patch experiment has been uploaded in[ tests/data/connection_test](https://github.com/ZeitgeberH/patchview/tree/main/tests/data/connection_test).
Documentation of the [analysis](https://patchview.readthedocs.io/en/latest/tutorials.html#monosynaptic-connection-detection) and [data](https://patchview.readthedocs.io/en/latest/installation.html#test-dataset) has been updated accordingly.
Thanks!
