### 2022-10-01 ZeitgeberH/patchview#10 @ZeitgeberH
Hi,
I've fixed "Sholl analysis" part. Please see documentation [here](https://patchview.readthedocs.io/en/latest/tutorials.html#morphological-data).
 You are right about the [Plotting somas for a group of neurons](https://patchview.readthedocs.io/en/latest/introduction.html#plotting-somas-for-a-group-of-neurons) functionality. It only works for reconstruction file with multiple cells in it (like this[ one](https://github.com/ZeitgeberH/patchview/blob/main/tests/data/test_multiSoma.ASC)). Likewise,  "Update Cell names"  is only useful under this  situation.
"Distance to Pia" only works if the[ ASC file](https://github.com/ZeitgeberH/patchview/blob/main/tests/data/test_multiSoma.ASC) actually has an entry named "Pia" .
Documentation of the [these functions](https://patchview.readthedocs.io/en/latest/tutorials.html#morphological-data) has been updated.
Thanks!
