### 2022-10-01 ZeitgeberH/patchview#11 @ZeitgeberH
I've updated the configure file location using appdirs. It works (see documentation [here](https://patchview.readthedocs.io/en/latest/installation.html#configuration)) great under Windows and Ubuntu.
Thanks for your suggestion.
