### 2022-10-01 c8226f0
write before read nbw
