### 2022-10-01 ZeitgeberH/patchview#8 @ZeitgeberH
Thanks for the catch. This has been fixed at [c8226f0](https://github.com/ZeitgeberH/patchview/commit/c8226f012f3b1864358138ae424ebb2574923f22)
