+++ patchview/patchview.py
@@ -964,6 +964,7 @@ class MainWindow(QtWidgets.QMainWindow):
         self.gpAction.setStatusTip("Analysis gap junction")
         self.gpAction.triggered.connect(self.gp_clicked)
         self.toolbar.addAction(self.gpAction)
+
         GPcorr_icon = os.path.join(patchview_dir, "Data", "icons", "GPcorr.png")
         self.gpAction2 = pg.QtGui.QAction(pg.QtGui.QIcon(GPcorr_icon), "correlations")
         # self.exportAction.setShortcut("Alt+e")
@@ -1049,35 +1050,17 @@ class MainWindow(QtWidgets.QMainWindow):
         self.fileMenu.addAction(self.resetAction)
 
         self.exitAction = pg.QtGui.QAction("&Exit")
-        #        self.saveFileAction.setShortcut("Ctrl+s")
         self.exitAction.triggered.connect(self.exit_clicked)
         self.fileMenu.addAction(self.exitAction)
 
-        self.AnalysisMenu = self.mbar.addMenu("&Analysis")
-        self.AnaAction1 = pg.QtGui.QAction("&Firing pattern analysis")
-        self.AnaAction1.setStatusTip(
-            "Select a firing pattern series to extract ephy features"
-        )
-        self.AnaAction1.triggered.connect(self.extraEphyAction_clicked)
-        self.AnalysisMenu.addAction(self.AnaAction1)
-
-        self.AnaAction4 = pg.QtGui.QAction("&Gap junction (Batch)")
-        # self.AnaAction4.triggered.connect(self.checkGapJunctionActionBatch_clicked)
-        self.AnaAction4.setStatusTip("Check gap connections in selected series")
-        self.AnalysisMenu.addAction(self.AnaAction4)
-
-        self.AnaAction5 = pg.QtGui.QAction("&Gap junction (Current tree)")
-        self.AnaAction5.triggered.connect(self.checkGapJunctionAction_clicked)
-        self.AnaAction5.setStatusTip("Check gap connections for current series")
-        self.AnalysisMenu.addAction(self.AnaAction5)
-
-        self.AnaAction2 = pg.QtGui.QAction("&Event detection")
-        self.AnaAction2.setStatusTip("Detect spontaneous EPSPs")
-        self.AnalysisMenu.addAction(self.AnaAction2)
-        self.AnaAction2.triggered.connect(self.eventDetectionAction_clicked)
+        ## TODO
+        # self.AnalysisMenu = self.mbar.addMenu("&Analysis")
+        # self.AnaAction4 = pg.QtGui.QAction("&Gap junction (Batch)")
+        # # self.AnaAction4.triggered.connect(self.checkGapJunctionActionBatch_clicked)
+        # self.AnaAction4.setStatusTip("Check gap connections in selected series")
+        # self.AnalysisMenu.addAction(self.AnaAction4)
 
         self.VisualMenu = self.mbar.addMenu("&Visualization")
-
         self.VisualAction2 = pg.QtGui.QAction("&Show averaged traces")
         self.VisualAction2.setShortcut("Ctrl+m")
         self.VisualAction2.setStatusTip("Plot averaged stimulated response")
@@ -1099,17 +1082,8 @@ class MainWindow(QtWidgets.QMainWindow):
         )  ## fine tune firing pattern plotting with user define parameters
         self.VisualMenu.addAction(self.VisualAction4)
 
-        self.VisualAction5 = pg.QtGui.QAction("&FP for all slice")
-        self.VisualAction5.setStatusTip(
-            "Selected traces for all cells in current slice"
-        )
-        self.VisualAction5.triggered.connect(
-            self.sliceFPView_plot
[…]
