+++ docs/installation.rst
@@ -189,7 +189,7 @@ of multiple neurons from a single brain slice.
 
 Tests
 -----------------
-Run pytest in the `tests`` folder 
+Run pytest in the `tests` folder 
 
 .. code-block:: console
 
