### 2022-10-03 ZeitgeberH/patchview#12 @ZeitgeberH
> * I do not quite understand the use of separators in the toolbar – are these placeholders for future functionality? Or just a way to logically group elements that belong together?

Toolbar is reduced to four items: mouse mode, loading slice image, calculating coupling ratio (CR), calculating correlation values (corr).
Separators are removed.

> * I was a bit confused by the fact that "Fire pattern" appears as:
>
>   1. A menu item (Analysis → Fire pattern analysis)

Removed.

>   2. A menu option (Analysis → Fire pattern batch analysis)

Removed.

>   3. A tab on the bottom left and on the top

Bottom left tap is for listing data with firing patterns. Top tap is for presenting figures and results.

>   4. A context-menu item in the tab on the bottom left ("Firing patterns (batch)")

This context menu is for batch processing firing pattern (when there are multiple series/cell firing patterns) loaded.

>   5. An entry in the toolbar ("Firing pattern")
>      → Maybe this could be reduced/unified/clarified a bit?

 Yes, This has been removed.

>
> * At a first glance, I did not immediately get the meaning of the separate "help tabs" that explain parameter settings (e.g. "Spike detection help" for "Spike detection" parameters), and they are not very readable, I think. Maybe you could consider adding the information as tooltips to the parameters directly?

Tooltips are added. Hovering over each value (the second column in each row) would show tooltip for that parameter.

> * The GUI contains a few typos that might be worth being corrected. For example, "Vislization" → "Visualization" and "Backgroud color" → "Background color" in the menu; "replacem value" → "replacement value" in the Event Detection/preprocessing parameters

Typo corrected.

Thanks!
