### 2022-10-03 d56655e
tooltips for spike detection parameters
