### 2022-10-03 540706a
coupling ratio test data
