### 2022-10-03 #4706 @ZeitgeberH
Hi @mstimberg Thanks very much for your suggestions.  I believe I have addressed all of the issues (details under each issue in the repo).  Additional test data, documentation and PyPI package has been updated.
