### 2022-10-05 ZeitgeberH/patchview#10 @mstimberg
I did a quick check with the provided files, and everything seems to work as intended now. Thanks.
