### 2022-10-05 #4706 @ZeitgeberH
> @ZeitgeberH do you have a version and DOI that corresponds to the JOSS submission (e.g., hosted on Zenodo)?

@jmschrei I just published current version on [Zenodo](https://zenodo.org/record/7149877) with DOI:
10.5281/zenodo.7149877
