### 2022-10-05 #4706 @jmschrei
I think you need a numeric format, e.g. https://github.com/openjournals/joss-reviews/issues/4680#issuecomment-1259594153
