### 2022-10-05 #4706 @ZeitgeberH
> I think you need a numeric format, e.g. [#4680 (comment)](https://github.com/openjournals/joss-reviews/issues/4680#issuecomment-1259594153)

Got it.
The tag is [v0.2.6](https://github.com/ZeitgeberH/patchview/releases/tag/v0.2.6)
