### 2022-10-09 #4706 @arfon
> @mstimberg, @janfreyberg – many thanks for your reviews here and to @ZeitgeberH for editing this submission! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨
>
> @jmschrei – your paper is now accepted and published in JOSS ⚡🚀💥

And of course I _actually_ mean, thanks for @jmschrei for editing this submission and @ZeitgeberH _your_ paper is now accepted and published in JOSS :-)
